# dist/js/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "v1.0.0",
        "title": "Swagger Restaurante Project",
        "description": "Pos Tech FIAP"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": ""
        }
    ],
    "paths": {
        "/product/categoria/{categoria}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "GetCategoria",
                "description": "Endpoint to get the list of specific product on category. Ex: Lanche, Combo, Sobremesa, Bebida",
                "parameters": [
                    {
                        "name": "categoria",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/product/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "GetID",
                "description": "Endpoint to get the specific product.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete",
                "description": "Endpoint to delete a product",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Update",
                "description": "Endpoint to update a product",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/product"
                            }
                        }
                    }
                }
            }
        },
        "/product/": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create",
                "description": "Endpoint to add a product",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/product"
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "GetAll",
                "description": "Endpoint to get all products",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/product/deactive/{id}": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Deactive",
                "description": "Endpoint to deactive a product",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Endpoint to get the specific user.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user/": {
            "post": {
                "tags": [
                    "User"
                ],
                "description": "Endpoint to add a user.",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/user"
                            }
                        }
                    }
                }
            }
        },
        "/cart/": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Create",
                "description": "Endpoint to create a cart",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/cart/user/{id}": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add a User",
                "description": "Endpoint to add a user to cart",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/cart/product/{id}": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add a Product",
                "description": "Endpoint to add a product to cart",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "product",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/cart/itens/{id}": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Personalize itens",
                "description": "Endpoint to personalize product itens",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "product",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "options",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/cart/{id}": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Resume",
                "description": "Endpoint to resume a cart",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/cart/close/{id}": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Close",
                "description": "Endpoint to close a cart",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/cart/pay/{id}": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Pay",
                "description": "Endpoint to pay a cart",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/cart/kitchen/{id}": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Send to Kitchen",
                "description": "Endpoint to send to kitchen a cart",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/cart/cancel/{id}": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Cancel",
                "description": "Endpoint to cancel a cart",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/order/receive/{id}": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Receive",
                "description": "Endpoint to receive a order",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/order/prepare/{id}": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Prepare",
                "description": "Endpoint to update status to prepare a order",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/order/estimate/{id}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Estimatte",
                "description": "Endpoint to calcute estimate of time from a order",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/order/update/ready/{id}": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Ready",
                "description": "Endpoint to update status to ready",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/order/update/delivered/{id}": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Delivery",
                "description": "Endpoint to update status to delivered",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/order/update/closed/{id}": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Close",
                "description": "Endpoint to update status to closed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/order/": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "GetAll Active",
                "description": "Endpoint to get all active orders",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "user": {
                "type": "object",
                "properties": {
                    "cpf": {
                        "type": "string",
                        "example": "000.000.000-00"
                    },
                    "name": {
                        "type": "string",
                        "example": "Jessica"
                    },
                    "email": {
                        "type": "string",
                        "example": "jessica.jessica@example.com"
                    }
                },
                "required": [
                    "cpf",
                    "name"
                ]
            },
            "product": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Sorvete Misto"
                    },
                    "options": {
                        "type": "string",
                        "example": "[Casquina, Massa Mista]"
                    },
                    "category": {
                        "type": "string",
                        "example": "sobremesa"
                    },
                    "price": {
                        "type": "string",
                        "example": "10"
                    },
                    "timeToPrepare": {
                        "type": "string",
                        "example": "2"
                    },
                    "status": {
                        "type": "string",
                        "example": "true"
                    }
                },
                "required": [
                    "name",
                    "options",
                    "category",
                    "price",
                    "timeToPrepare",
                    "status"
                ]
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}